requirements.txt
requests~=2.26.0
awscli
flask~=2.0.2
selenium~=4.0.0
Pillow~=8.3.2